refactor: move responses from pipeline.rs --> responses.rs
