+++ sgl-router/src/routers/grpc/mod.rs
@@ -7,6 +7,7 @@ pub mod conversions;
 pub mod pd_router;
 pub mod pipeline;
 pub mod processing;
+pub mod responses;
 pub mod router;
 pub mod streaming;
 pub mod utils;
